--- BrianFiles/ImportTesting/Main.py
@@ -1,5 +1,0 @@
-from MainHelp.Generic import *
-
-
-
-print("This is Main")
