--- BrianFiles/ImportTesting/MainHelp/Generic.py
@@ -1,3 +1,0 @@
-
-
-print("This is generic")
