--- BrianFiles/ImportTesting/TestHelp/Helper.py
@@ -1,3 +1,0 @@
-from MainHelp.Generic import *
-
-print("This is Helper")
