--- BrianFiles/ImportTesting/TestMain.py
@@ -1,3 +1,0 @@
-import TestHelp.Helper
-
-print("This is TestMain")
