+++ BrianFiles/Parsers/JsonNodeParser.py
@@ -1,5 +1,5 @@
-from Node import *
-from NodeParser import NodeParser
+from Containers.Node import *
+from Parsers.NodeParser import NodeParser
 
 '''
 The JsonNodeParser class is responsible for constructing a node conatiner from a .json file
